+++ frontend/src/components/prompts/Move.vue
@@ -12,7 +12,7 @@
       class="card-action"
       style="display: flex; align-items: center; justify-content: space-between"
     >
-      <template v-if="user.perm.create">
+      <template v-if="user.perm.modify">
         <button
           class="button button--flat"
           @click="$refs.fileList.createDir()"