+++ frontend/src/components/settings/Permissions.vue
@@ -2,35 +2,13 @@
   <div>
     <h3>{{ $t("settings.permissions") }}</h3>
     <p class="small">{{ $t("settings.permissionsHelp") }}</p>
-
     <p>
       <input type="checkbox" v-model="admin" />
       {{ $t("settings.administrator") }}
     </p>
-
-    <p>
-      <input type="checkbox" :disabled="admin" v-model="perm.create" />
-      {{ $t("settings.perm.create") }}
-    </p>
-    <p>
-      <input type="checkbox" :disabled="admin" v-model="perm.delete" />
-      {{ $t("settings.perm.delete") }}
-    </p>
-    <p>
-      <input type="checkbox" :disabled="admin" v-model="perm.download" />
-      {{ $t("settings.perm.download") }}
-    </p>
     <p>
       <input type="checkbox" :disabled="admin" v-model="perm.modify" />
       {{ $t("settings.perm.modify") }}
-    </p>
-    <p v-if="isExecEnabled">
-      <input type="checkbox" :disabled="admin" v-model="perm.execute" />
-      {{ $t("settings.perm.execute") }}
-    </p>
-    <p>
-      <input type="checkbox" :disabled="admin" v-model="perm.rename" />
-      {{ $t("settings.perm.rename") }}
     </p>
     <p>
       <input type="checkbox" :disabled="admin" v-model="perm.share" />